Refactor docker stack
I had nothing but trouble getting multisite to work with apache. When ever I finished the install I would get either a database conection error or I would get stuck on the login page in a redirect loop. Everything seems to be working with the current stack. On to testing it on the production server.