Refactored wp.conf for WordPress Mulitisite
subdirectory install. I have decided to go with this form of network as it makes more sense to as an api. I am also concidering if .work would be more fitting for this project. Need to set up a way to manage the domain in one spot rather than having to change it in several places.