Set up nginx to use env variables
Had to use a shell script to process all the templates for each site. So that included refactoring the docker/nginx to have a ./templates and ./sites.
The big problem I ran into was when a .conf file uses many variables. Apart of that shell script was to help preserve those variables by adding DOLLAR var that just holds '$'. then I needed to find each nginx var in the file and add  where the normal $ would be. Will need to set up a conditional in the static.conf to check if requests for svgs are coming from folio or blog and so set the CORS resonse header appropriately.