Add set up to filter post archive.
I originally thought I would need to reorder the post archive, but it as I thought it over all I really need to do is hide all but the selected categorey. I have made a style rule selecting the all the cards in .active-filter and hides them except those with .filter. Next I need to write a function that adds the active filter class to post excerpts' container. Then this function needs to cycle through all the post and add filter to the class list.