Post archive filter working as expected.
The filter menu needs a couple more things. A way to turn off all filters. When the filter is activated, a user can make it so none of the page's content is render. Rather it would be great to have it turn off the all filters if none are active.

Also updated portfolio post to have more specific excerpts and updated the JS to use the posts' excerpts rather than the content. This will eventually allow me to set up a single page to give users more information about what I did for a specific client or project.